% Generated by build script
\name{artic_artists}
\alias{artic_artists}
\title{Browse artists at the Art Institute of Chicago}
\usage{
artic_artists()
}
\arguments{
\item{limit}{Number of artists per page (default 20)}
\item{page}{Page number (default 1)}
}
\value{tibble: id, title, birth_date, death_date, description}
\description{Browse artists at the Art Institute of Chicago}
