% Generated by build script
\name{artic_exhibitions}
\alias{artic_exhibitions}
\title{Search Art Institute exhibitions}
\usage{
artic_exhibitions()
}
\arguments{
\item{query}{Search query (optional)}
\item{limit}{Number of results (default 20)}
}
\value{tibble: id, title, status, aic_start_at, aic_end_at}
\description{Search Art Institute exhibitions}
