% Generated by build script
\name{artic_search}
\alias{artic_search}
\title{Search Art Institute of Chicago artworks}
\usage{
artic_search()
}
\arguments{
\item{query}{Search query string}
\item{limit}{Maximum results (default 20, max 100)}
}
\value{tibble: id, title, artist, date_display, medium, department,}
\description{Search Art Institute of Chicago artworks}
